Switch to cross-compiler, update bug list, add more PCI functions

+++ source/kernel/gfx/terminal.c
@@ -650,7 +650,7 @@ void printf_putchar(const char *format, void(*put_method)(char), va_list args) {
                     // %x or %X is for hexadecimal
                     case 'X':
                     case 'x': {
-                        unsigned int c = va_arg(args, unsigned int); // We need to get the next integer (it will be converted to hexadecimal when we use itoa.) from the parameters.
+                        unsigned long c = va_arg(args, unsigned long); // We need to get the next integer (it will be converted to hexadecimal when we use itoa.) from the parameters.
                         char str[32] = {0}; // String buffer where the hexadecimal will go.
 
                         // itoa() will convert the integer to base-16 or hex.
